Completely changed Makefile to a TDD-oriented one.
It was taken from another source of my own I already had.

This will allow the project to depend on the selected
tests in order for the main app to successfully build.

For now, TDD is disabled. To have a module compilation
success depend on its test success, move it from OBJS_NO_TDD
to OBJS_TDD.

To write a test for 'module.o' you create 'module.t.c'.

When testing, Make will create module.t, which will be executed.

If it succeeds, module.ts (for 'test succeed') will be touched.